@@ -4,9 +4,9 @@ export default class PluginLoader {
   constructor(private readonly resolvers: Array<Resolver>) {}
 
   async load(source: string): Promise<object> {
-    for (let resolver of this.resolvers) {
+    for (const resolver of this.resolvers) {
       if (await resolver.canResolve(source)) {
-        let resolvedPath = await resolver.resolve(source);
+        const resolvedPath = await resolver.resolve(source);
         return require(resolvedPath);
       }
     }
