@@ -18,7 +18,7 @@ export class NetworkLoadError extends Error {
 export default class NetworkResolver implements Resolver {
   async canResolve(source: string): Promise<boolean> {
     try {
-      let url = new URL(source);
+      const url = new URL(source);
       return url.protocol === 'http:' || url.protocol === 'https:';
     } catch {
       return false;
@@ -26,13 +26,13 @@ export default class NetworkResolver implements Resolver {
   }
 
   async resolve(source: string): Promise<string> {
-    let response = await get(source, { followRedirect: true });
+    const response = await get(source, { followRedirect: true });
 
     if (response.statusCode !== 200) {
       throw new NetworkLoadError(response);
     }
 
-    let filename = tmp({ postfix: '.js' });
+    const filename = tmp({ postfix: '.js' });
     await writeFile(filename, response.body);
     return filename;
   }
