@@ -7,11 +7,11 @@ export default async function copyFixturesInto(
   fixture: string,
   destination: string
 ): Promise<string> {
-  let fixturePath = join('test/fixtures', fixture);
+  const fixturePath = join('test/fixtures', fixture);
 
-  for (let file of iterateSources([fixturePath], null, () => false)) {
-    let relativePath = relative(fixturePath, file.path);
-    let destinationPath = join(destination, relativePath);
+  for (const file of iterateSources([fixturePath], null, () => false)) {
+    const relativePath = relative(fixturePath, file.path);
+    const destinationPath = join(destination, relativePath);
     await mkdirp(dirname(destinationPath));
     await writeFile(destinationPath, file.content);
   }
