@@ -4,7 +4,7 @@ import getTemporaryFilePath from './getTemporaryFilePath';
 export default async function createTemporaryDirectory(
   path: string
 ): Promise<string> {
-  let fullPath = getTemporaryFilePath(path);
+  const fullPath = getTemporaryFilePath(path);
   await mkdirp(fullPath);
   return fullPath;
 }
