@@ -132,7 +132,7 @@ module.exports = {
   // testLocationInResults: false,
 
   // The glob patterns Jest uses to detect test files
-  testMatch: ['**/__tests__/*.+(ts|tsx|js)'],
+  // testMatch: ['**/__tests__/*.+(ts|tsx|js)'],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
   // testPathIgnorePatterns: [
@@ -140,7 +140,7 @@ module.exports = {
   // ],
 
   // The regexp pattern Jest uses to detect test files
-  // testRegex: "",
+  testRegex: '/__tests__/(test|.*\\.test)\\.ts$',
 
   // This option allows the use of a custom results processor
   // testResultsProcessor: null,
