@@ -1,7 +1,7 @@
 import Matcher from './Matcher';
 
 export class AnythingMatcher<T> extends Matcher<T> {
-  // eslint-disable-next-line typescript/no-unused-vars
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   matchValue(value: unknown): value is T {
     return true;
   }
